重构代码，增加统一的ak/sk获取、日志定义
