+++ src/server.py
@@ -1,37 +1,28 @@
 from mcp.server.fastmcp import FastMCP
-
 import logging
-
 from tools import ecs_tools
+from tools import obs_tools
+from tools import utils
+
 
 logger = logging.getLogger(__name__)
 
-# Create an MCP server
-mcp = FastMCP("HuaweiCloud-ECS-ops-mcp-server")  
-@mcp.tool()
 
-@click.command()
-
-@click.option(
-    "--transport",
-    type=click.Choice(["stdio", "sse"]),
-    default="stdio",
-    help="Transport type",
-)
 def main(transport: str):
+    utils.setup_logging()
+
     # Create an MCP server
-    mcp = FastMCP("alibaba-cloud-ops-mcp-server")
+    mcp = FastMCP("huawei-cloud-mcp-server")
     for tool in ecs_tools.tools:
         mcp.add_tool(tool)
-    
-    api_tools.create_api_tools(mcp, config)
+
+    for tool in obs_tools.tools:
+        mcp.add_tool(tool)
 
     # Initialize and run the server
     logger.debug(f'mcp server is running on {transport} mode.')
     mcp.run(transport=transport)
-    logger.debug(f"mcp server is running on {transport} mode.")
-    mcp.run(transport=transport)
 
 
 if __name__ == "__main__":
-    main()
+    main("stdio")
